server/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
Flask==2.3.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==0.28.1
packaging==24.2
psycopg2-binary==2.9.7
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
